C++ Solution
            Submission Statistics:
            Question #: 58
            Language: C++
            Runtime: 0 ms
            Runtime percentile: 100.00
            Memory: 6.468 MB
            Memory percentile: 84.55

+++ 58. Length of Last Word/Solution.cpp
@@ -4,8 +4,8 @@ public:
         int length = 0;
         bool wasSpace = false;
 
-        for (int i=0; i<s.length(); ++i){
-            if (static_cast<int>(s[i]) != 32){ // not a space
+        for (int i=0; i < s.length(); ++i){
+            if (s[i] != 32){ // not a space
                 if (wasSpace){
                     length = 0;
                     wasSpace = false;
